pyproject.toml
[project]
name = "ergomate-homeassistant"
version = "0.1.0"
description = "Home Assistant integration for Ergomate standing desks via Bluetooth Low Energy"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "bleak>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.2",
    "pytest-homeassistant-custom-component>=0.13.109",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = [".git"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.2",
    "pytest-homeassistant-custom-component>=0.13.109",
    "pytest-asyncio>=0.23.0",
]
